+++ src/collective/contact/plonegroup/upgrades/upgrades.py
@@ -75,13 +75,15 @@ def v6(context):
     logger.info("Migrate to v6")
     functions = api.portal.get_registry_record(name=FUNCTIONS_REGISTRY)
     res = []
+    # XXX we intentionally directly edit the existing "function"
+    # so it does not break in the detectContactPlonegroupChange method while looking
+    # for 'enabled' in old_functions
     for function in functions:
         if 'fct_management' not in function:
             function['fct_management'] = False
-            res.append(function)
-        else:
-            # already migrated
-            return
+        if 'enabled' not in function:
+            function['enabled'] = True
+        res.append(function)
     api.portal.set_registry_record(FUNCTIONS_REGISTRY, res)
     setup = api.portal.get_tool('portal_setup')
     setup.runImportStepFromProfile('profile-collective.contact.plonegroup:default', 'plone.app.registry')